quantity compoent
