+++ basket/src/components/BasketItem.tsx
@@ -1,12 +1,7 @@
 import React from "react";
 import { List, Space, Button, Typography, Avatar } from "antd";
-import {
-  PlusOutlined,
-  MinusOutlined,
-  DeleteOutlined,
-  CheckCircleOutlined,
-} from "@ant-design/icons";
 import type { CartItem } from "@library/types";
+import { QuantityControl } from "./QuantityControl";
 
 const { Text } = Typography;
 
@@ -62,48 +57,5 @@ export const BasketItem: React.FC<BasketItemProps> = ({
         }
       />
     </List.Item>
-  );
-};
-
-interface QuantityControlProps {
-  itemId: number;
-  quantity: number;
-  onUpdateQuantity: (id: number, quantity: number) => void;
-}
-
-export const QuantityControl: React.FC<QuantityControlProps> = ({
-  itemId,
-  quantity,
-  onUpdateQuantity,
-}) => {
-  const quantityStyle: React.CSSProperties = {
-    width: 40,
-    textAlign: "center",
-    borderTop: "1px solid #d9d9d9",
-    borderBottom: "1px solid #d9d9d9",
-    padding: "4px 0",
-    display: "inline-flex",
-    alignItems: "center",
-    justifyContent: "center",
-    height: "24px",
-  };
-
-  return (
-    <Space.Compact style={{ marginTop: "8px" }}>
-      <Button
-        size="small"
-        icon={<MinusOutlined />}
-        onClick={() => onUpdateQuantity(itemId, quantity - 1)}
-        disabled={quantity <= 1}
-      />
-      <div style={quantityStyle}>
-        <Text>{quantity}</Text>
-      </div>
-      <Button
-        size="small"
-        icon={<PlusOutlined />}
-        onClick={() => onUpdateQuantity(itemId, quantity + 1)}
-      />
-    </Space.Compact>
   );
 };
